internet: (fixes #1282) Fix integer overflow in TcpPrrRecovery
In the Proportional Rate Reduction branch of TcpPrrRecovery::DoRecovery,
the expression "m_prrDelivered * tcb->m_ssThresh * 1.0" evaluated the
product of the two uint32_t operands before promoting to double. After a
few dozen segments into recovery the product exceeds 2^32 and wraps
around, yielding a far-too-small congestion window and degraded
throughput under loss.

Cast m_prrDelivered to uint64_t so the multiplication is performed in
64-bit arithmetic. Add a regression test that drives the computation past
the uint32_t range and asserts the resulting cwnd matches the 64-bit
reference value.

+++ src/internet/model/tcp-prr-recovery.cc
@@ -86,8 +86,13 @@ TcpPrrRecovery::DoRecovery(Ptr<TcpSocketState> tcb, uint32_t deliveredBytes, boo
     if (tcb->m_bytesInFlight > tcb->m_ssThresh)
     {
         // Proportional Rate Reductions
-        sendCount =
-            std::ceil(m_prrDelivered * tcb->m_ssThresh * 1.0 / m_recoveryFlightSize) - m_prrOut;
+        // The product m_prrDelivered * m_ssThresh must be evaluated in 64-bit
+        // arithmetic: both operands are uint32_t and their product overflows a
+        // uint32_t after only a few dozen segments into recovery, which would
+        // wrap around and shrink the congestion window incorrectly (issue #1282).
+        sendCount = std::ceil(static_cast<uint64_t>(m_prrDelivered) * tcb->m_ssThresh * 1.0 /
+                              m_recoveryFlightSize) -
+                    m_prrOut;
     }
     else
     {
